backend/requirements.txt
# OfferU - Backend Dependencies
fastapi>=0.128.0
uvicorn[standard]>=0.30.0
sse-starlette>=3.3.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings>=2.5.0
eval_type_backport>=0.2.0  # Python 3.9 compat for X | Y in Pydantic models
httpx>=0.27.0
openai>=2.30.0
python-multipart==0.0.9
weasyprint>=60.0
reportlab>=4.0.0
PyMuPDF>=1.24.0
playwright>=1.51.0
jinja2>=3.1.0
python-docx>=1.1.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
python-jobspy>=1.1.80
jieba>=0.42.0
mcp>=1.27.0
